requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.2
chromadb==0.5.0
ollama==0.2.1
praw==7.7.1
google-analytics-data==0.18.7
google-auth==2.29.0